re-style movie list and movie item components
