+++ src/components/MovieList/MovieItem.js
@@ -1,59 +1,75 @@
 import React, { useContext } from "react";
 import { nominateStore, ACTIONS } from "../Store/nominateStore";
-import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
+
+// @material ui
 import CardActions from "@material-ui/core/CardActions";
+import Button from "@material-ui/core/Button";
+import { makeStyles } from "@material-ui/core/styles";
+import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import CardMedia from "@material-ui/core/CardMedia";
-import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
-import { makeStyles } from "@material-ui/core/styles";
+import { MoviesDataWrapper } from "./Style";
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
   root: {
-    maxWidth: 345,
+    display: "flex",
+    width: "600px",
   },
-  media: {
-    height: 140,
+  details: {
+    display: "flex",
+    flexDirection: "column",
+    width: 400,
+    justifyContent: "center",
   },
-});
+  content: {
+    flex: "1 0 auto",
+  },
+  cover: {
+    width: 200,
+  },
+}));
 function MovieItem({ movie }) {
   const { dispatch } = useContext(nominateStore);
-  const { Year, Type, Title, Poster } = movie;
-  function handleclick() {
+
+  function handleClick() {
     dispatch({
       type: ACTIONS.ADD_NOMINATE,
       data: movie,
     });
   }
 
+  const { Year, Type, Title, Poster } = movie;
+
   // @material-ui useStyles
   const classes = useStyles();
+  // const theme = useTheme();
+
   return (
-    <Card className={classes.root}>
-      <CardActionArea>
-        <CardMedia
-          className={classes.media}
-          image={Poster}
-          alt={Title}
-          title={Title}
-        />
-        <CardContent>
-          <Typography gutterBottom variant="h5" component="h2">
-            {Title}
-          </Typography>
-          <Typography variant="body2" color="textSecondary" component="p">
-            <p>Year: {Year} </p>
-            <p>Type: {Type}</p>
-          </Typography>
-        </CardContent>
-      </CardActionArea>
-      <CardActions>
-        <Button size="small" color="primary" onClick={() => handleclick()}>
-          Nominate
-        </Button>
-      </CardActions>
-    </Card>
+    <MoviesDataWrapper>
+      <Card className={classes.root}>
+        <CardMedia className={classes.cover} image={Poster} title={Title} />
+        <div className={classes.details}>
+          <CardContent className={classes.content}>
+            <Typography component="h5" variant="h5">
+              {Title}
+            </Typography>
+
+            <Typography component="h5" variant="h5">
+              Type: {Type}
+            </Typography>
+            <Typography component="h5" variant="h5">
+              Year: {Year}
+            </Typography>
+          </CardContent>
+          <CardActions>
+            <Button size="large" color="primary" onClick={() => handleClick()}>
+              Nominate
+            </Button>
+          </CardActions>
+        </div>
+      </Card>
+    </MoviesDataWrapper>
   );
 }
