+++ src/components/MovieList/Style.js
@@ -2,7 +2,6 @@ import styled from "styled-components";
 
 export const MoviesDataWrapper = styled.div`
   display: flex;
-  margin: 10px;
+  margin: 15px;
   justify-content: center;
-
 `;
